+++ src/pages/Home/TutorialSection/index.tsx
@@ -5,7 +5,7 @@ interface TutorialProps {
   content: TutorialLink[];
 }
 
-export const Tutorial = ({ content }: TutorialProps) => (
+export const TutorialSection = ({ content }: TutorialProps) => (
   <section className="flex flex-col gap-4 bg-yellow-bright-01 px-8 py-10">
     {content.map((item) => (
       <div
